Update resource_fetcher.py
Changes to the user agent are done to reduce the number of exploits that hackers can try to abuse. The user agent is upgraded to the most up to date version of Microsoft Edge. Before it was running Google Chrome 131 which has extreme amount of exploits that are patched in the latest release. And Edge is now more secure than Google Chrome with having security features that Chrome never ever had at all.