Add files via upload
Here I add a check whether the PC uses Legacy BIOS and if yes, the script checks if the CPU is first gen i3/i5/i7 or Xeon "Nehalem" and if yes, it informs the user that with these PCs on Legacy BIOS OpenCore is unstable unless a third party BIOS has been flashed and that Clover is recommended instead. And gives the user options whether to continue experimenting or exit OpCore-Simplify and continue with Clover instead. Another thing is to say that OCLP may disable macOS security funcitons but it doesn't do it always - only when needed. And of course, the large portion of the changes are to fix a couple of vulnerabilities:
- an attacker could attach over a gigabyte JSON that is actually invalid, that could lead to crashing the entire script or malformed efi - mitigated by limiting the size to 100MB
-a JSON file could execute files outside the OpCore-Simplify's folder which could lead to malware execution - mitigated by removing the ability the JSON file to execute malicious code
- an attacker could upload an empty or JSON files that only execute commands only and no hardware report - which can lead to downloading payloads instead of generating a valid EFI. This is mitigated by not allowing JSON files to execute commands.
- an attacker could delete libraries outside the workspace of OpCore-Simplify - which could lead to corrupting the OS completely or introducing OS vulnerabilities

Signed-off-by: bobbystanev-gif <bobby.stanev@gmail.com>