Add files via upload
The updater now can troubleshoot any issues by itself if found or guide the user how to fix those issues. This will save time for the users who are using this script. Like if they use Windows 8.1, on that they need to upgrade or if they use Windows 10 and for some reason it doesn't work, it can update or upgrade Windows to Windows 11 to fix this issue and will check if it meets the minimum requirements if it decides it's time to upgrade. There might be some bugs to fix though and will take time to look on those.

Signed-off-by: bobbystanev-gif <bobby.stanev@gmail.com>